requirements.txt
# Lambda-optimized requirements - minimal footprint
fastapi==0.104.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
mangum==0.17.0

# Excluded for Lambda (not needed):
# - uvicorn (Lambda handles ASGI)
# - alembic (migrations run separately)
# - python-dotenv (Lambda uses env vars directly)
